Sample `deadline` from a exponential distribution
This commit adds a reveal time which is when agents learn about their
payment obligation. Deadlines are then sampled from and exponential dist
using [`rand_distr`](https://docs.rs/rand_distr/latest/rand_distr/).
And refactors the prng factory to use pcg which implements `RngCore`.
Lastly, amounts are also sampled from an exponential dist.

+++ src/main.rs
@@ -8,9 +8,9 @@ struct Args {
     #[arg(short, long)]
     payment_obligations: usize,
     #[arg(short, long)]
-    time_steps: usize,
+    time_steps: u64,
     #[arg(short, long)]
-    block_interval: usize,
+    block_interval: u64,
     #[arg(short, long)]
     seed: Option<u64>,
 }
